backend/run.py
from flask import Flask
from app.agents.ecom_agent import init_app
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

app = Flask(__name__)

# Initialize the ecom agent
init_app(app)

if __name__ == '__main__':
    # Run the Flask app
    app.run(host='0.0.0.0', port=5000, debug=True) 