backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
flask-jwt-extended==4.6.0
pydantic==1.10.13
pymongo==4.6.1
python-jose==3.3.0 